Refactor package hierarchy
