+++ src/main/java/seedu/duke/ui/UserInterfaceConstants.java
@@ -1,9 +1,9 @@
-package seedu.duke.main;
+package seedu.duke.ui;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Constants {
+public class UserInterfaceConstants {
     public static final String INTRO_MESSAGE = "Welcome to LogJob!";
     public static final String HELP_MESSAGE = "Hi this is a template help message found in the Constants class!";
     public static final String LOGO = """
@@ -19,7 +19,7 @@ public class Constants {
     public static final List<String> TABLE_HEADERS =
             List.of("ID", "Company", "Job Title", "Status", "Date of Application");
 
-    public static final ArrayList<String> TABLE_HEADER_ARRAYLIST = new ArrayList<>(Constants.TABLE_HEADERS);
+    public static final ArrayList<String> TABLE_HEADER_ARRAYLIST = new ArrayList<>(UserInterfaceConstants.TABLE_HEADERS);
 
     public static final String LINE_BREAK = "----------------------------------------";
 }
